feat: additional provides
